Create image download script
