+++ packages/cms/package.json
@@ -12,10 +12,7 @@
     "test:ci": "yarn run test",
     "compile": "echo 'Placeholder to support yarn compile in repo root, until we use TS'",
     "lint": "echo '@TODO add eslint'",
-    "assets": "run-s assets:download assets:image-resize",
-    "assets:download": "bash scripts/assets-download.sh",
-    "assets:image-resize": "ts-node scripts/images-resize.ts",
-    "assets2": "ts-node scripts/assets.ts --resolveJsonModule"
+    "download-assets": "ts-node scripts/download-assets.ts"
   },
   "keywords": [
     "sanity"
@@ -39,11 +36,13 @@
     "sanity-plugin-tabs": "^2.0.1"
   },
   "devDependencies": {
-    "sharp": "^0.27.0",
-    "npm-run-all": "^4.1.5",
     "@sanity/cli": "^2.2.6",
     "@types/download": "^6.2.4",
+    "@types/prompts": "^2.0.9",
     "@types/sharp": "^0.27.1",
-    "download": "^8.0.0"
+    "download": "^8.0.0",
+    "npm-run-all": "^4.1.5",
+    "prompts": "^2.4.0",
+    "sharp": "^0.27.0"
   }
 }
